Tuvung/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai